Move handling of pending_cmd to parse_input_event
Previously `pending_cmd' was a field in State.t and `Game.execute' was
responsible for updating it. By taking `pending_cmd' out of State.t and
making it an argument to `parse_input_event', we decouple the State
module from the Cmd module and allow several functions (besides execute)
that should have been in Cmd in the first place to be moved back to Cmd.
We also get to remove a lot of boilerplate code from Game that was
basically parsing input events twice, although much of this refactoring
has yet to be done.