Make it compile
There's some stuff here from like literally a year ago and I don't remember what
I was trying to accomplish but I'm going to commit it anyway because this is the
version that compiles.

I had to comment out a lot of code to get it to compile, which will be addressed
by the subsequent refactoring.